+++ tests.py
@@ -6,30 +6,12 @@ import main
 
 sys.path.insert(0, "testing")
 import mock_test_file
+import test_helpers
 
 main.set_running_tests()
 
 # Add Test Name Here
 tests = ["basic_load_config()", "basic_test()"]
-
-# Helpers
-def pre_tests():
-    os.system("cp savedata.toml savedatacopy.toml")
-    os.system("cp -r testing testing_backup")
-
-def cleanup(exit_code):
-    os.system("mv savedatacopy.toml savedata.toml")
-    os.system("mv testing_backup testing")
-    quit(exit_code)
-    
-def base64_encode(img_path):
-    with open(img_path, "rb") as i:
-        contents = i.read()
-    return base64.b64encode(contents)
-
-def write_config(config_file_contents):
-    with open("savedata.toml", "w") as f:
-        f.write(config_file_contents)
 
 # Tests
 def basic_load_config():
@@ -38,7 +20,7 @@ def basic_load_config():
     directory = "."
     supervision_level = 0
     """
-    write_config(config_contents)
+    test_helpers.write_config(config_contents)
     config = main.load_config()
     expected = {"repo": "test", "directory": ".", "supervision_level": 0}
     if config == expected:
@@ -52,7 +34,7 @@ def basic_test():
     directory = "."
     supervision_level = 0
     """
-    write_config(config_contents)
+    test_helpers.write_config(config_contents)
     config = main.load_config()
     expected = {"repo": "test", "directory": ".", "supervision_level": 0}
     if config != expected:
@@ -75,11 +57,11 @@ def basic_test():
         # Test should now pass after images are updated
         return False
 
-pre_tests()
+test_helpers.pre_tests()
 fail = False
 for i in tests:
     if eval(i) == False:
         print("Test "+i+" failed!")
         fail = True
 
-cleanup(1 if fail else 0)
+test_helpers.cleanup(1 if fail else 0)
